feat: toggle Center Spread by making an HTTP request to the receiver
This commit implements SurroundParameterSettingsRepository which now
uses a newly added Retrofit2 API client to fetch and set the value of
Center Spread on the receiver.

This means that the app can now be used to toggle the setting on an
actual real-life receiver.

This commit adds an e2e integration test which sends actual HTTP
requests to a local fake receiver.